+++ src/textual/widgets/_data_table.py
@@ -1096,9 +1096,10 @@ class DataTable(ScrollView, Generic[CellType], can_focus=True):
             elif self.cursor_type == "column":
                 self.refresh_column(old_coordinate.column)
                 self._highlight_column(new_coordinate.column)
-            # If the coordinate was changed via `move_cursor`, give priority to its
-            # scrolling because it may be animated.
-            self.call_after_refresh(self._scroll_cursor_into_view)
+            if abs(new_coordinate.row - old_coordinate.row) <= 1:
+                self._scroll_cursor_into_view()
+            else:
+                self.call_after_refresh(self._scroll_cursor_into_view)
 
     def move_cursor(
         self,
@@ -1137,7 +1138,11 @@ class DataTable(ScrollView, Generic[CellType], can_focus=True):
         # of rows then tried to immediately move the cursor.
         # We do this before setting `cursor_coordinate` because its watcher will also
         # schedule a call to `_scroll_cursor_into_view` without optionally animating.
-        self.call_after_refresh(self._scroll_cursor_into_view, animate=animate)
+        # self.call_after_refresh(self._scroll_cursor_into_view, animate=animate)
+        if self._require_update_dimensions:
+            self.call_after_refresh(self._scroll_cursor_into_view, animate=animate)
+        else:
+            self._scroll_cursor_into_view(animate=animate)
 
         self.cursor_coordinate = destination